requirements.txt
tensorflow==2.12.0
matplotlib==3.1.3
mediapipe
numpy
opencv-python
Pillow
websocket-client
rel